requirements.txt
django==3.2.16
django-environ==0.10.0
django-cors-headers==3.14.0
djangorestframework==3.14.0

argon2-cffi==21.3.0
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.5.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0

Pillow==9.5.0
django-storages==1.13.2
django-ckeditor==6.5.1
psycopg2==2.9.6
python-slugify==6.1.2
requests==2.28.1
crispy-bootstrap5==0.7
django-crispy-forms==2.0
mysqlclient==2.1.1